timing: report elapsed user time rather than wall time
This commit changes the application's reporting of elapsed time to
measure user time rather than wall time in order to reduce sources
of noise in evaluating runtime.

For output is now of the form:

  ```
  Processing time (wall): 123.456 s
  Processing time (user): 123.456 s
  ```

Since the application performs some operations through subprocesses,
it is necessary to address a lack of getrusage(RUSAGE_CHILDREN)
equivalent on the Windows platform.  This is handled by wrapping
system() in order to account the child's process times when the
child has finished executing.




git-svn-id: http://wg11.sc29.org/svn/repos/MPEG-I/Part5-PointCloudCompression/TM/TMC2/trunk@1328 94298a81-5874-47c9-aab8-bfb24faeed7f